+++ packages/core/package.json
@@ -42,6 +42,7 @@
     "dev": "unbuild --stub"
   },
   "dependencies": {
-    "@magic-color/shared": "workspace:*"
+    "@magic-color/shared": "workspace:*",
+    "@magic-color/transformer": "workspace:*"
   }
 }
