+++ packages/core/src/basic/utils.ts
@@ -1,8 +1,6 @@
 import type { ColorObject, Colors, ColorType, Opacity } from '@magic-color/core'
 import { isArray, isNumber, isObject, isString } from '@magic-color/shared'
-import { isHex, isHsb, isHsl, isKeyword, isRgb, parseHex, parseHsb, parseHsl, parseKeyword, parseLab, parseRgb } from '@magic-color/transformer'
-
-export const SupportTypes = ['rgb', 'hex', 'hsl', 'hsb', 'lab']
+import { isHex, isHsb, isHsl, isKeyword, isRgb, parseHex, parseHsb, parseHsl, parseKeyword, parseLab, parseRgb, SupportTypes } from '@magic-color/transformer'
 
 export function alphaToString(alpha: Opacity, toHex = false): string {
   return toHex
@@ -154,8 +152,4 @@ export function resolveArgs<T extends ColorType = 'rgb'>(...args: any[]): [Color
 
     return [_values, _type, _alpha]
   }
-}
-
-export function valid(color: string) {
-  return guessType(color) !== undefined
 }
