+++ packages/core/src/index.ts
@@ -1,5 +1,3 @@
-export * from './colors/colorbrewer'
-export * from './colors/uno'
-export * from './colors/w3cx11'
-
+export * from './basic'
 export * from './types'
+export * from '@magic-color/transformer'
