--- packages/transform/src/index.ts
@@ -1,7 +1,0 @@
-export * from './transform/hex'
-export * from './transform/hsb'
-export * from './transform/hsl'
-export * from './transform/keywords'
-
-export * from './transform/lab'
-export * from './transform/rgb'
