+++ packages/transformer/build.config.ts
@@ -3,6 +3,10 @@ import { defineBuildConfig } from 'unbuild'
 export default defineBuildConfig({
   entries: [
     'src/index',
+    {
+      input: 'src/colors/index',
+      name: 'colors',
+    },
   ],
   declaration: true,
   clean: true,
