+++ packages/transformer/src/meta/hsb.ts
@@ -1,4 +1,4 @@
-import type { HsbColor, HslColor, LabColor, RgbColor } from '@magic-color/core'
+import type { HsbColor, HslColor, LabColor, RgbColor } from '../types'
 import { rgbToHex, rgbToLab } from './rgb'
 
 const hsbRegex = /^hsb\((\d+)\s*,\s*(\d+)%\s*,\s*(\d+)%\)$/
