+++ packages/transformer/src/meta/keywords.ts
@@ -1,18 +1,18 @@
-import type { HexColor } from '@magic-color/core'
-import { colorbrewer, unoColors, w3cx11 } from '@magic-color/core'
+import type { HexColor } from '../types'
 import { isObject } from '@magic-color/shared'
+import { Colorbrewer, UnoColors, W3cx11 } from '../colors'
 
 import { parseHex } from './hex'
 
 export function isKeyword(color: string): boolean {
-  return color in unoColors || color in colorbrewer || color in w3cx11
+  return color in UnoColors || color in Colorbrewer || color in W3cx11
 }
 
 export function parseKeyword(color: string) {
   let hex
   let c: any
 
-  c = (unoColors as any)[color]
+  c = (UnoColors as any)[color]
   if (c) {
     if (isObject(c))
       hex = c.DEFAULT
@@ -23,13 +23,13 @@ export function parseKeyword(color: string) {
     return parseHex(hex as HexColor)
   }
 
-  c = (colorbrewer as any)[color]
+  c = (Colorbrewer as any)[color]
   if (c) {
     hex = c[3]
     return parseHex(hex)
   }
 
-  c = (w3cx11 as any)[color]
+  c = (W3cx11 as any)[color]
   if (c)
     return parseHex(c)
 }
