+++ packages/transformer/src/meta/rgb.ts
